requirement.txt
streamlit
fpdf
google-generativeai
speechrecognition
pydub
python-dotenv
